qa/requirements.txt
pytest-bdd==7.3.0
requests==2.31.0
selenium==4.14.0
pytest-xdist==3.3.1
allure-pytest-bdd==2.13.2
